Keep only the desktop implementation for OpenLink module
