+++ jpro-openlink/src/main/java/one/jpro/platform/openlink/OpenLink.java
@@ -1,47 +1,50 @@
 package one.jpro.platform.openlink;
 
-import com.jpro.webapi.WebAPI;
 import javafx.stage.Stage;
-import one.jpro.platform.openlink.impl.DesktopOpenLinkImpl;
-import one.jpro.platform.openlink.impl.JProOpenLinkImpl;
+import one.jpro.platform.openlink.util.PlatformUtils;
 import org.jetbrains.annotations.NotNull;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
 import java.net.URL;
+import java.util.Optional;
 
 /**
- * OpenLink interface for opening URLs.
+ * Launches the default browser of the platform as a separate application process. The browser
+ * will be opened with the provided URL.
  *
  * @author Besmir Beqiri
  */
-public interface OpenLink {
+public final class OpenLink {
 
-    /**
-     * Creates an instance of OpenLink based on the given application stage.
-     *
-     * @param stage the application stage
-     * @return an instance of OpenLink
-     */
-    static OpenLink create(Stage stage) {
-        if (WebAPI.isBrowser()) {
-            WebAPI webAPI = WebAPI.getWebAPI(stage);
-            return new JProOpenLinkImpl(webAPI);
-        } else {
-            return new DesktopOpenLinkImpl();
-        }
-    }
+    private static final Logger logger = LoggerFactory.getLogger(OpenLink.class);
 
     /**
      * Opens the given URL in the browser.
      *
-     * @param uri the URL to open
+     * @param url the URL to open
      */
-    void openURL(@NotNull URL url);
-
+    public void openURL(@NotNull URL url) {
+        openURL(url.toString());
+    }
 
     /**
      * Opens the given URL string in the browser.
      *
-     * @param uri the URL string to open
+     * @param url the URL string to open
      */
-    void openURL(@NotNull String url);
+    public void openURL(@NotNull String url) {
+        try {
+            if (PlatformUtils.isMac()) {
+                Runtime.getRuntime().exec("open " + url);
+            } else if (PlatformUtils.isWindows()) {
+                Runtime.getRuntime().exec("start \"" + url + "\"");
+            } else if (PlatformUtils.isLinux()) {
+                Runtime.getRuntime().exec("xdg-open " + url);
+            }
+        } catch (IOException ex) {
+            logger.error("Unable to open the browser!", ex);
+        }
+    }
 }
